finish sectionpage and bug fix to header
