+++ app/assets/javascripts/section/SectionHeader.js
@@ -1,3 +1,4 @@
+import React from 'react';
 const Routes = window.shared.Routes;
 
 class SectionHeader extends React.Component {
